LayoutTests/platform/gtk/imported/w3c/web-platform-tests/html/semantics/popovers/popover-focus-2.tentative-expected.txt

Button1  Button2  Invoker1  Button3  Button4
Invoker  after
Show popover

FAIL Popover focus navigation assert_true: popover1 should be invoked by invoker1 expected true got false
FAIL Circular reference tab navigation assert_equals: Step 2 expected Element node <button id="circular1" autofocus="" popoverhidetarget="po... but got Element node <button id="circular4">after</button>
FAIL Popover focus returns when popover is hidden by invoker assert_true: popover should be invoked by invoker expected true got false
